feat: add new skills and update save functionality
- Added new skills: Breathing, Swimming, Walking, and Gorilla.
- Updated save functionality to include online mode and real-time options.
- Modified settings to use i8 for fog and lighting.
- Added rand dependency for random time generation.
- Updated ToDo list to reflect completed tasks.
- Removed unused time.cpp and time.hpp files.

+++ src/settings.rs
@@ -5,8 +5,8 @@ use std::path::Path;
 
 #[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
 pub struct Settings {
-	pub fog: i32,
-	pub lighting: i32,
+	pub fog: i8,
+	pub lighting: i8,
 	pub sound: bool,
 	pub developer_mode: bool,
 	pub verbose_logging: bool,
